Move frame/series to eager folder
