+++ py-polars/polars/__init__.py
@@ -1,13 +1,12 @@
 # flake8: noqa
 from .datatypes import *
-from .frame import *
+from .eager import *
 from .functions import *
 from .lazy import *
-from .series import *
 
 # during docs building the binary code is not yet available
 try:
-    from .frame import version
+    from .polars import version
 
     __version__ = version()
 except ImportError:
