+++ py-polars/polars/eager/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .frame import *
+from .series import *
